refactor(radix tree): moves radix tree in subfolder
