+++ src/trees/radix/node.ts
@@ -1,5 +1,5 @@
-import type { Nullable } from "../types/index.js";
-import { uniqueId } from "../utils.js";
+import type { Nullable } from "../../types/index.js";
+import { uniqueId } from "../../utils.js";
 
 export interface RadixNode {
   id: string;
