+++ tests/radix.test.ts
@@ -1,12 +1,13 @@
 import t from "tap";
-import { create as createNode } from "../src/radix-tree/node.js";
+import { create as createNode } from "../src/trees/radix/node.js";
 import {
   contains as radixContains,
   find as radixFind,
   insert as radixInsert,
   removeDocumentByWord as radixRemoveDocumentByWord,
   removeWord as radixRemoveWord,
-} from "../src/radix-tree/radix.js";
+} from "../src/trees/radix/index.js";
+
 const phrases = [
   { id: "1", doc: "the quick, brown fox" },
   { id: "2", doc: "jumps over the lazy dog" },
